backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "python-dotenv>=1.2.2",
    "waitress>=3.0.2",
]
